migrate to new tempus client

+++ src/TempusHub.Core/Utilities/HrefHelper.cs
@@ -5,10 +5,10 @@ namespace TempusHub.Core.Utilities;
 public static class HrefHelper
 {
     public static string GetMapInfoPath(string mapName) => "/map/" + mapName;
-    public static string GetPlayerInfoPath(int playerId) => GetPlayerInfoPath(playerId.ToString(CultureInfo.InvariantCulture));
+    public static string GetPlayerInfoPath(long playerId) => GetPlayerInfoPath(playerId.ToString(CultureInfo.InvariantCulture));
     public static string GetPlayerInfoPath(string playerId) => "/player/" + playerId;
     public static string GetServerInfoPath(string server) => "/server/" + server;
-    public static string GetDemoInfoPath(int demoId) => "/demo/" + demoId;
-    public static string GetRunInfoPath(int runId) => "/run/" + runId;
+    public static string GetDemoInfoPath(long demoId) => "/demo/" + demoId;
+    public static string GetRunInfoPath(long runId) => "/run/" + runId;
 
 }
